Add GET handler for config API and fetch protocols in settings
