+++ frontend/src/components/server-config-stepper.tsx
@@ -6,15 +6,18 @@ import useSWR from 'swr';
 import axios from 'axios';
 
 import {
+  Tooltip,
+  TooltipContent,
+  TooltipProvider,
+  TooltipTrigger,
+} from '@/components/ui/tooltip';
+import {
   Card,
   CardContent,
   CardDescription,
   CardHeader,
   CardTitle,
 } from '@/components/ui/card';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
-
 import {
   Select,
   SelectContent,
@@ -22,13 +25,8 @@ import {
   SelectTrigger,
   SelectValue,
 } from '@/components/ui/select';
-
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipProvider,
-  TooltipTrigger,
-} from '@/components/ui/tooltip';
+import { Input } from '@/components/ui/input';
+import { Label } from '@/components/ui/label';
 import { Separator } from '@/components/ui/separator';
 
 import {
@@ -42,43 +40,8 @@ import {
 import { z } from 'zod';
 import { useRouter } from 'next/navigation';
 
-interface Service {
-  id: string;
-  name: string;
-  port: number;
-}
-
-interface ConfigData {
-  general: {
-    protocol: string;
-    tick_time: number;
-    flag_ttl: number;
-    start_time: string;
-    end_time: string;
-  };
-  flagChecker: {
-    url_flag_checker: string;
-    team_token: string;
-    submit_flag_checker_time: number;
-    max_flag_batch_size: number;
-  };
-  flagInfo: {
-    regex_flag: string;
-    url_flag_ids: string;
-  };
-  services: Service[];
-  teams: {
-    range_ip_teams: number;
-    nop_team: number;
-    my_team_id: number;
-    format_ip_teams: string;
-  };
-}
-
-interface Protocol {
-  value: string;
-  label: string;
-}
+import { Service, ConfigData, Protocol } from '@/lib/types'; // Adjust the import path as needed
+import { fetcher } from '@/lib/fetchers';
 
 const initialConfig: ConfigData = {
   general: {
@@ -105,12 +68,6 @@ const initialConfig: ConfigData = {
     my_team_id: 1,
     format_ip_teams: '10.60.{}.1',
   },
-};
-
-const fetcher = async (url: string) => {
-  const response = await axios.get(url);
-  console.log(`Fetching data from ${url}:`, response.data);
-  return response.data;
 };
 
 const steps = [
@@ -907,9 +864,9 @@ export function ServerConfigStepper() {
               </div>
               <Input
                 id="nop-team"
-                placeholder="1"
+                placeholder="0"
                 type="number"
-                value={config.teams.nop_team || ''}
+                value={config.teams.nop_team || 0}
                 onChange={e =>
                   updateConfig('teams', 'nop_team', e.target.value)
                 }
