+++ frontend/src/lib/types.ts
@@ -11,3 +11,41 @@ export type Flag = {
   port_service: number;
   team_id: number;
 };
+
+export type Protocol = {
+  value: string;
+  label: string;
+};
+
+export type Service = {
+  id: string;
+  name: string;
+  port: number;
+};
+
+export type ConfigData = {
+  general: {
+    protocol: string;
+    tick_time: number;
+    flag_ttl: number;
+    start_time: string;
+    end_time: string;
+  };
+  flagChecker: {
+    url_flag_checker: string;
+    team_token: string;
+    submit_flag_checker_time: number;
+    max_flag_batch_size: number;
+  };
+  flagInfo: {
+    regex_flag: string;
+    url_flag_ids: string;
+  };
+  services: Service[];
+  teams: {
+    range_ip_teams: number;
+    nop_team: number;
+    my_team_id: number;
+    format_ip_teams: string;
+  };
+};
